LayoutTests/fast/block/float/margin-top-changes.html

<!DOCTYPE html>
<style>
body
{
  margin: 0;
  padding: 0;
}
.div
{
  width: 100px;
  height: 50px;
}
#float
{
  float: left;
}
#inline
{
  display: inline-block;
  background-color:green;
}
#positioned
{
  position: absolute;
  background-color: red;
  z-index: -1;
}
</style>
<div id="positioned" class="div"></div>
<div id="float" class="div"></div>
<div id="test">
  <div id="inline" class="div"></div>
</div>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
  document.body.offsetTop;
  float.style["margin-top"] = "-50px";
    test(() => {
        assert_equals(document.getElementById("inline").offsetLeft, 0);
    }, "crbug.com/711938: Sibling elements notice when a float no longer overlaps due to margin top changing.");
</script>
